Use an intermediate shell script as the condor executable
When using a virtualenv'd python, I found that it would try and
transfer over the symlink to the python executable, which would
result in immediate 'import site' crashes.  Now it is wrapped
in a shell script (which gets cleaned up along w/ the stdxxx).

Addtionally, the input pickle is transferred using the
'transfer_input_files' scheme.  I had a hard time getting it to
run in the correct AFS directory, even with initialdir=os.getcwd()